Move icon_theme related stuff in separate file
* icon_theme_helper.py: file contains icon_theme helper functions
* Submakefile: add icon_theme_helper to gmoccapy modules
* notification.py and gmoccapy.py: Use new helper functions
